requirements.txt
fastapi
uvicorn
pydantic
requests
diskcache
structlog
python-dotenv
docxtpl
streamlit
pandas
numpy
pyyaml
pytest
